rm some exports
