+++ packages/core/src/http/getHTTPStatusCode.ts
@@ -1,11 +1,10 @@
 import { TRPCError } from '../error/TRPCError';
-import { invert } from '../internals/invert';
-import { TRPC_ERROR_CODES_BY_KEY, TRPCResponse } from '../rpc';
+import {
+  TRPC_ERROR_CODES_BY_KEY,
+  TRPC_ERROR_CODES_BY_NUMBER,
+  TRPCResponse,
+} from '../rpc';
 
-export const TRPC_ERROR_CODES_BY_NUMBER = invert(TRPC_ERROR_CODES_BY_KEY);
-type ValueOf<TType> = TType[keyof TType];
-
-export type TRPC_ERROR_CODE_NUMBER = ValueOf<typeof TRPC_ERROR_CODES_BY_KEY>;
 const JSONRPC2_TO_HTTP_CODE: Record<
   keyof typeof TRPC_ERROR_CODES_BY_KEY,
   number
