+++ packages/core/src/http/types.ts
@@ -1,5 +1,6 @@
-import { AnyRouter, inferRouterContext, ProcedureType } from '..';
-import { BaseHandlerOptions } from '../internals/types';
+import { TRPCError } from '../error/TRPCError';
+import { AnyRouter } from '../router';
+import { inferRouterContext, ProcedureType } from '../types';
 import { HTTPHeaders, ResponseMetaFn } from './internals/types';
 
 export interface HTTPRequest {
@@ -49,3 +50,27 @@ export type TRPCRequestInfo = {
  */
 export type ResolveHTTPRequestOptionsContextFn<TRouter extends AnyRouter> =
   (opts: { info: TRPCRequestInfo }) => Promise<inferRouterContext<TRouter>>;
+/**
+ * Base interface for any response handler
+ * @internal
+ */
+
+export interface BaseHandlerOptions<TRouter extends AnyRouter, TRequest> {
+  onError?: OnErrorFunction<TRouter, TRequest>;
+  batching?: {
+    enabled: boolean;
+  };
+  router: TRouter;
+}
+/**
+ * @internal
+ */
+
+export type OnErrorFunction<TRouter extends AnyRouter, TRequest> = (opts: {
+  error: TRPCError;
+  type: ProcedureType | 'unknown';
+  path: string | undefined;
+  req: TRequest;
+  input: unknown;
+  ctx: inferRouterContext<TRouter> | undefined;
+}) => void;
