--- packages/core/src/internals/types.ts
@@ -1,27 +1,0 @@
-import { inferRouterContext, ProcedureType } from '../';
-import { TRPCError } from '../error/TRPCError';
-import { AnyRouter } from '../router';
-
-/**
- * Base interface for any response handler
- * @internal
- */
-export interface BaseHandlerOptions<TRouter extends AnyRouter, TRequest> {
-  onError?: OnErrorFunction<TRouter, TRequest>;
-  batching?: {
-    enabled: boolean;
-  };
-  router: TRouter;
-}
-
-/**
- * @internal
- */
-export type OnErrorFunction<TRouter extends AnyRouter, TRequest> = (opts: {
-  error: TRPCError;
-  type: ProcedureType | 'unknown';
-  path: string | undefined;
-  req: TRequest;
-  input: unknown;
-  ctx: inferRouterContext<TRouter> | undefined;
-}) => void;
