+++ packages/core/turbo.json
@@ -3,13 +3,7 @@
   "extends": ["//"],
   "pipeline": {
     "codegen-entrypoints": {
-      "outputs": [
-        "package.json",
-        "http/**",
-        "internals/**",
-        "observable/**",
-        "rpc/**"
-      ]
+      "outputs": ["package.json", "http/**", "observable/**", "rpc/**"]
     }
   }
 }
