+++ packages/react-query/src/unstableDoNotImportThis.ts
@@ -4,3 +4,5 @@
  * If you need to import anything from here, please open an issue.
  */
 export * from '@trpc/core';
+export * from '@trpc/core/http';
+export * from '@trpc/core/rpc';
