+++ packages/server/package.json
@@ -113,6 +113,11 @@
       "import": "./dist/shared.mjs",
       "require": "./dist/shared.js",
       "default": "./dist/shared.js"
+    },
+    "./unstableDoNotImportThis": {
+      "import": "./dist/unstableDoNotImportThis.mjs",
+      "require": "./dist/unstableDoNotImportThis.js",
+      "default": "./dist/unstableDoNotImportThis.js"
     }
   },
   "files": [
@@ -125,6 +130,7 @@
     "observable",
     "rpc",
     "shared",
+    "unstableDoNotImportThis",
     "!**/*.test.*"
   ],
   "publishConfig": {
