+++ packages/server/src/unstableDoNotImportThis.ts
@@ -1,0 +1,8 @@
+/**
+ * This file is here to make TypeScript happy.
+ *
+ * If you need to import anything from here, please open an issue.
+ */
+export * from '@trpc/core';
+export * from '@trpc/core/http';
+export * from '@trpc/core/rpc';
