+++ packages/server/turbo.json
@@ -9,7 +9,8 @@
         "http/**",
         "observable/**",
         "rpc/**",
-        "shared/**"
+        "shared/**",
+        "unstableDoNotImportThis/**"
       ]
     }
   }
